# Guardrails Module Fixes Walkthrough
I have resolved the compilation errors in the `wayang-guardrails-core` module by standardizing the SPI implementations and fixing visibility and import issues.

## Changes Made

### SPI Compliance and Interface Implementations
- **Detectors**: Updated [BiasDetector](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/BiasDetector.java#9-123), [HallucinationDetector](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/HallucinationDetector.java#9-194), [PIIDetector](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/PIIDetector.java#10-92), and [ToxicityDetector](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/ToxicityDetector.java#9-105) to correctly implement the `WayangPlugin` interface by adding missing [name()](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/BiasDetector.java#17-21), [id()](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/BiasDetector.java#12-16), [version()](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/ToxicityDetector.java#26-30), and [description()](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/PIIDetector.java#28-32) methods.
- **Node Provider**: Refactored [GuardrailNodeProvider](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/node/GuardrailNodeProvider.java#10-78) to align with the [NodeProvider](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/core/wayang-plugin-spi/src/main/java/tech/kayys/wayang/plugin/spi/node/NodeProvider.java#49-56) SPI, removing the non-existent `NodeCategory` and ensuring correct imports.

### Visibility and Package Fixes
- **Policy Severity**: Changed `PolicySeverity` from a package-private enum to a `public enum` to resolve access errors in [GuardrailPolicy](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/GuardrailPolicy.java#6-14).
- **Finding Class**: Ensured the [Finding](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/Finding.java#3-10) class and its usages are consistent across the detector package.

### Standardized Usage
- **NodeContext**: Unified the usage of [NodeContext](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/NodeContext.java#5-13) to point to the correct `tech.kayys.wayang.guardrails.NodeContext` package across [PolicyEngine](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/policy/PolicyEngine.java#23-165), [GuardrailPolicyPlugin](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/plugin/GuardrailPolicyPlugin.java#14-35), and [PolicyBasedGuardrailsEngine](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/policy/PolicyBasedGuardrailsEngine.java#12-76).
- **ExecutionResult**: Fixed [ContentRedactor](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/ContentRedactor.java#10-54) to use the [ExecutionResult](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/ExecutionResult.java#5-9) defined within the guardrails module instead of a non-existent external package.
- **CheckPhase**: Standardized `CheckPhase` usage across the module, particularly in [GuardrailNodeExecutor](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/executor/GuardrailNodeExecutor.java#29-123).

## Verification Results

The changes address all 10+ compilation errors originally reported. The components are now correctly aligned with the Gamelan SDK 2.x and Wayang SPI patterns.

### Key Files Modified:
- [PolicySeverity.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/policy/PolicySeverity.java)
- [GuardrailPolicyPlugin.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/plugin/GuardrailPolicyPlugin.java)
- [PolicyEngine.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/policy/PolicyEngine.java)
- [PolicyBasedGuardrailsEngine.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/policy/PolicyBasedGuardrailsEngine.java)
- [ContentRedactor.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/ContentRedactor.java)
- [GuardrailNodeProvider.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/node/GuardrailNodeProvider.java)
- [BiasDetector.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/BiasDetector.java)
- [HallucinationDetector.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/HallucinationDetector.java)
- [PIIDetector.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/PIIDetector.java)
- [ToxicityDetector.java](file:///Users/bhangun/Workspace/workkayys/Products/Wayang/wayang-platform/wayang/executors/guardrails/guardrails-core/src/main/java/tech/kayys/wayang/guardrails/detector/ToxicityDetector.java)

+++ executors/memory/memory-core/src/main/java/tech/kayys/wayang/memory/context/HierarchicalMemoryManager.java
@@ -280,17 +280,17 @@ public class HierarchicalMemoryManager {
     }
 
     // Utility methods
-    private double calculateSimilarity(List<Double> vec1, List<Double> vec2) {
+    private double calculateSimilarity(List<Float> vec1, List<Float> vec2) {
         if (vec1 == null || vec2 == null || vec1.isEmpty() || vec2.isEmpty()) return 0.0;
         if (vec1.size() != vec2.size()) return 0.0;
-        
+
         double dotProduct = 0.0, norm1 = 0.0, norm2 = 0.0;
         for (int i = 0; i < vec1.size(); i++) {
             dotProduct += vec1.get(i) * vec2.get(i);
             norm1 += vec1.get(i) * vec1.get(i);
             norm2 += vec2.get(i) * vec2.get(i);
         }
-        
+
         return dotProduct / (Math.sqrt(norm1) * Math.sqrt(norm2));
     }
